remove: fix retry logic
The original logic properly identifies when the node to be removed has changed, and resets `p` so it points to `head` again. However, it misses a call to `continue` to jump back to the beginning of the loop. Failing to do so causes `p` to be set to a different value once again before going back to the start of the loop.